backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
supabase>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=21.0.0
python-jose[cryptography]>=3.3.0
pyjwt>=2.6.0
psycopg2-binary>=2.9.1